Moved DueDateLabel to its own file
